check: report number of failed test cases and skipped test cases
As a part of the Blktests CI preparation, improve the test result
reporting. Record failed test cases in the file named "failures" in the
results output directory. Print the count and the list of failed test
cases at the end of test runs. Do the same for skipped test cases using
the file named "skips".

The output will be like this:

  $ sudo ./check meta
  meta/001 (do nothing)                                        [passed]
      runtime  0.000s  ...  0.000s
  [...]
  meta/019 (x=1 y=1 z=1) (combine three set_conditions() hooks) [passed]
      runtime  0.000s  ...  0.000s
  4 failures: meta/003 meta/005 meta/006 meta/009
  2 skips: meta/007 meta/014

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>